nvim: change plugin manager to paq
packer.nvim stopped being maintained in mid-2023. It still works, but
obviously it's better to use maintained software than unmaintained
software.

I tried moving to lazy.nvim the other day, but it was a disaster,
requiring way too much work.

paq, on the other hand, is very much a drop-in replacement. The only
missing feature is specifying plugins by tags, but there was only one
that used that.