Cleanup inventory files
It has been bothering me for a while that group vars belong to two
different dimensions - one is OS-based ("Debian") and other is
function-based ("work"). This change finally cleans it up - now it's
function exclusively ("work" and "private").

To facilitate this change, split inventory files. Now it's expected to
use `-i`, `--inventory` flag instead of `-l`, `--limit`. README is
updated with a relevant note.